fix(*): add test for SQL injection in repository

+++ modules/engine/cstl-engine-admin/src/main/java/org/constellation/admin/UserBusiness.java
@@ -106,7 +106,7 @@ public class UserBusiness implements IUserBusiness {
     }
 
     @Override
-    public int countUser() {
+    public long countUser() {
         return userRepository.countUser();
     }
